Switch to absolute timestamps on the API
When I first designed this thing, it seemed like a good idea to have it return last seen values
relative to now. This turns out to be a terrible idea, so it now returns unix timestamps

Note: This is a breaking API change, but no one will care.

+++ apitypes/acnode.go
@@ -6,13 +6,13 @@ type ACNode struct {
 	MqttName string `json:"mqttName",omitempty`
 	Type string `json:"nodeType"`
 
-	// how many seconds ago was the node last seen?
-	LastSeen int `json:"LastSeen"`
-	LastSeenMQTT int `json:"LastSeenMQTT"`
-	LastSeenAPI int `json:"LastSeenAPI"`
+	// What was the timestamp this node was last seen at?
+	LastSeen int64 `json:"LastSeen"`
+	LastSeenMQTT int64 `json:"LastSeenMQTT"`
+	LastSeenAPI int64 `json:"LastSeenAPI"`
 
 	// When did we last see a START message for it?
-	LastStarted int `json:"LastStarted"`
+	LastStarted int64 `json:"LastStarted"`
 
 	MemFree int `json:"MemFree",omitempty`
 	MemUsed int `json:"MemUsed",omitempty"`
